+++ sbin/dataimport-02-import-tracks-to-import-database.sh
@@ -2,6 +2,7 @@
 # exit on error
 set -e
 CWD=$(pwd)
+SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 function dofail {
     cd $CWD
     printf '%s\n' "$1" >&2  ## Send message to stderr. Exclude >&2 if you don't want it that way.
@@ -14,7 +15,7 @@ if [ "$#" -ne 1 ]; then
     exit 1
 fi
 IMPORTKEY=$1
-./setImportDirectory.sh $IMPORTKEY
+${SCRIPTPATH}/setImportDirectory.sh $IMPORTKEY
 
 echo "OPEN: Do you want to import track to import-database?"
 select yn in "Yes" "No"; do
@@ -26,11 +27,11 @@ done
 echo "start - import track to import-database"
 
 echo "now: configure linux vars: run configure-environment.bash"
-source configure-environment.bash
+source ${SCRIPTPATH}/configure-environment.bash
 
 echo "now: initialize import-database (sqlite)"
 cd ${MYTB}
-node_modules/.bin/db-migrate up --migrations-dir migrations/mytbdb --config config/db-migrate-database.json --env mytbdb_import_sqlite3
+node_modules/.bin/db-migrate up --migrations-dir migrations/mytbdb --config ${CONFIG_BASEDIR}db-migrate-database.json --env mytbdb_import_sqlite3
 cd $CWD
 
 echo "YOUR TODO: start facetcache for import-database in separate shell: 'cd ${MYTB} && npm run backend-start-server-managed-facetcache-dev-import && cd $CWD'"
@@ -45,7 +46,7 @@ echo "now: load import-files"
 if [ -f "${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-images.json" ]; then
   echo "now: load image-import-file"
   cd ${MYTB}
-  node dist/backend/serverAdmin.js --debug --command loadTourDoc  -c config/backend.import.json -f ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-images.json
+  node dist/backend/serverAdmin.js --debug --command loadTourDoc  -c ${CONFIG_BASEDIR}backend.import.json -f ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-images.json
   mv  ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-images.json  ${MYTB_IMPORT_MEDIADIR}import/DONE-mytbdb_import-import-images.json
   rm  ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-images.tmp || echo ""
   cd $CWD
@@ -65,7 +66,7 @@ fi
 if [ -f "${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-videos.json" ]; then
   echo "now: load video-import-file"
   cd ${MYTB}
-  node dist/backend/serverAdmin.js --debug --command loadTourDoc  -c config/backend.import.json -f ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-videos.json
+  node dist/backend/serverAdmin.js --debug --command loadTourDoc  -c ${CONFIG_BASEDIR}backend.import.json -f ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-videos.json
   mv  ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-videos.json  ${MYTB_IMPORT_MEDIADIR}import/DONE-mytbdb_import-import-videos.json
   rm  ${MYTB_IMPORT_MEDIADIR}import/mytbdb_import-import-videos.tmp || echo ""
   cd $CWD
@@ -86,7 +87,7 @@ echo "OPTIONAL: read image-dates"
 echo "OPEN: Do you want to read the image-dates?"
 select yn in "Yes" "No"; do
     case $yn in
-        Yes ) cd ${MYTB} && node dist/backend/serverAdmin.js --command imageManager --action readImageDates -c config/backend.import.json; break;;
+        Yes ) cd ${MYTB} && node dist/backend/serverAdmin.js --command imageManager --action readImageDates -c ${CONFIG_BASEDIR}backend.import.json; break;;
         No) break;;
     esac
 done
@@ -95,7 +96,7 @@ echo "OPTIONAL: read video-dates"
 echo "OPEN: Do you want to read the video-dates?"
 select yn in "Yes" "No"; do
     case $yn in
-        Yes ) cd ${MYTB} && node dist/backend/serverAdmin.js --command imageManager --action readVideoDates -c config/backend.import.json; break;;
+        Yes ) cd ${MYTB} && node dist/backend/serverAdmin.js --command imageManager --action readVideoDates -c ${CONFIG_BASEDIR}backend.import.json; break;;
         No) break;;
     esac
 done
@@ -107,9 +108,9 @@ cd $CWD
 
 echo "now: create scaled video-copies"
 cd ${MYTB}
-node dist/backend/serverAdmin.js --command mediaManager --action generateVideoScreenshotFromMediaDir --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import//video_screenshot/ --debug true
-node dist/backend/serverAdmin.js --command mediaManager --action generateVideoPreviewFromMediaDir --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import/video_thumbnail/ --debug true
-node dist/backend/serverAdmin.js --command mediaManager --action scaleVideosFromMediaDirToMP4 --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import/video_x600/ --debug true
+node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.import.json --command mediaManager --action generateVideoScreenshotFromMediaDir --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import//video_screenshot/ --debug true
+node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.import.json --command mediaManager --action generateVideoPreviewFromMediaDir --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import/video_thumbnail/ --debug true
+node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.import.json --command mediaManager --action scaleVideosFromMediaDirToMP4 --importDir ${MYTB_IMPORT_MEDIADIR}import/video_full/ --outputDir ${MYTB_IMPORT_MEDIADIR}import/video_x600/ --debug true
 cd $CWD
 
 echo "OPTIONAL YOUR TODO: fix image/track-date if needed (via gui or cli and run script manually)"