+++ sbin/setImportDirectory.sh
@@ -2,6 +2,7 @@
 # exit on error
 set -e
 CWD=$(pwd)
+SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 function dofail {
     cd $CWD
     printf '%s\n' "$1" >&2  ## Send message to stderr. Exclude >&2 if you don't want it that way.
@@ -16,7 +17,7 @@ fi
 IMPORTKEY=$1
 
 echo "now: configure linux vars: run configure-environment.bash"
-source configure-environment.bash
+source ${SCRIPTPATH}/configure-environment.bash
 
 if [ ! -d "${MYTB_IMPORT_MEDIADIR}${IMPORTKEY}" ]; then
   echo "FATAL: import-directory '${MYTB_IMPORT_MEDIADIR}${IMPORTKEY} must exist"