+++ src/backend/commands/media-manager.command.ts
@@ -9,7 +9,13 @@ import {MediaManagerModule} from '@dps/mycms-server-commons/dist/media-commons/m
 import {CommonMediaManagerCommand} from '@dps/mycms-server-commons/dist/backend-commons/commands/common-media-manager.command';
 import {AbstractCommand} from '@dps/mycms-server-commons/dist/backend-commons/commands/abstract.command';
 import {TourDocFileUtils} from '../shared/tdoc-commons/services/tdoc-file.utils';
-import {FileSystemDBSyncType, ProcessingOptions} from '@dps/mycms-server-commons/dist/backend-commons/modules/cdoc-media-manager.module';
+import {FileSystemDBSyncType} from '@dps/mycms-server-commons/dist/backend-commons/modules/cdoc-media-manager.module';
+import {ProcessingOptions} from '@dps/mycms-commons/dist/search-commons/services/cdoc-search.service';
+import {TourDocMediaFileExportManager} from '../modules/tdoc-mediafile-export.service';
+import {MediaExportResolutionProfiles, TourDocExportService} from '../modules/tdoc-export.service';
+import {TourDocServerPlaylistService, TourDocServerPlaylistServiceConfig} from '../modules/tdoc-serverplaylist.service';
+import {TourDocMediaFileImportManager} from '../modules/tdoc-mediafile-import.service';
+import {MediaExportProcessingOptions} from '@dps/mycms-server-commons/dist/backend-commons/modules/cdoc-mediafile-export.service';
 
 export class MediaManagerCommand implements AbstractCommand {
     public process(argv): Promise<any> {
@@ -25,8 +31,22 @@ export class MediaManagerCommand implements AbstractCommand {
         if (writable) {
             dataService.setWritable(true);
         }
+        const playlistConfig: TourDocServerPlaylistServiceConfig = {
+            audioBaseUrl: backendConfig.playlistExportAudioBaseUrl,
+            imageBaseUrl: backendConfig.playlistExportImageBaseUrl,
+            videoBaseUrl: backendConfig.playlistExportVideoBaseUrl,
+            useAudioAssetStoreUrls: backendConfig.playlistExportUseAudioAssetStoreUrls,
+            useImageAssetStoreUrls: backendConfig.playlistExportUseImageAssetStoreUrls,
+            useVideoAssetStoreUrls: backendConfig.playlistExportUseVideoAssetStoreUrls
+        };
+        const playlistService = new TourDocServerPlaylistService(playlistConfig);
+        const tourDocMediaFileExportManager = new TourDocMediaFileExportManager(backendConfig.apiRoutePicturesStaticDir, playlistService);
+        const tourDocExportManager = new TourDocExportService(backendConfig, dataService, playlistService, tourDocMediaFileExportManager,
+            new TourDocAdapterResponseMapper(backendConfig));
+        const tourDocMediaFileImportManager = new TourDocMediaFileImportManager(backendConfig, dataService);
         const mediaManagerModule = new MediaManagerModule(backendConfig.imageMagicAppPath, os.tmpdir());
-        const tdocManagerModule = new TourDocMediaManagerModule(backendConfig, dataService, mediaManagerModule);
+        const tdocManagerModule = new TourDocMediaManagerModule(backendConfig, dataService, mediaManagerModule, tourDocExportManager,
+            tourDocMediaFileImportManager);
         const commonMediadManagerCommand = new CommonMediaManagerCommand(backendConfig);
 
         let promise: Promise<any>;
@@ -36,6 +56,8 @@ export class MediaManagerCommand implements AbstractCommand {
             parallel: Number.parseInt(argv['parallel'], 10)
         };
         const pageNum = Number.parseInt(argv['pageNum'], 10);
+        const playlists = argv['playlists'];
+        const personalRateOverall = argv['personalRateOverall'];
         switch (action) {
             case 'readImageDates':
                 processingOptions.parallel = Number.isInteger(processingOptions.parallel) ? processingOptions.parallel : 1;
@@ -62,6 +84,57 @@ export class MediaManagerCommand implements AbstractCommand {
 
                 promise = tdocManagerModule.scaleImagesToDefaultWidth(searchForm, processingOptions);
 
+                break;
+            case 'exportImageFiles':
+            case 'exportVideoFiles':
+                const exportDir = argv['exportDir'];
+                if (exportDir === undefined) {
+                    console.error(action + ' missing parameter - usage: --exportDir EXPORTDIR', argv);
+                    promise = utils.reject(action + ' missing parameter - usage: --exportDir EXPORTDIR [-force true/false]');
+                    return promise;
+                }
+                const directoryProfile = argv['directoryProfile'];
+                if (directoryProfile === undefined) {
+                    console.error(action + ' missing parameter - usage: --directoryProfile directoryProfile', argv);
+                    promise = utils.reject(action + ' missing parameter - usage: --directoryProfile directoryProfile');
+                    return promise;
+                }
+
+                const fileNameProfile = argv['fileNameProfile'];
+                if (fileNameProfile === undefined) {
+                    console.error(action + ' missing parameter - usage: --fileNameProfile fileNameProfile', argv);
+                    promise = utils.reject(action + ' missing parameter - usage: --fileNameProfile fileNameProfile');
+                    return promise;
+                }
+
+                const resolutionProfile = argv['resolutionProfile'];
+                if (resolutionProfile === undefined || !Object.keys(MediaExportResolutionProfiles).includes(resolutionProfile)) {
+                    console.error(action + ' missing parameter - usage: --resolutionProfile {' +
+                        Object.keys(MediaExportResolutionProfiles).join(', ') + '}', argv);
+                    promise = utils.reject(action + ' missing parameter - usage: --resolutionProfile {' +
+                        Object.keys(MediaExportResolutionProfiles).join(', ') + '}');
+                    return promise;
+                }
+
+                const exportName = argv['exportName'];
+
+                processingOptions.parallel = Number.isInteger(processingOptions.parallel) ? processingOptions.parallel : 1;
+                searchForm = new TourDocSearchForm({
+                    type: action === 'exportVideoFiles' ? 'video' : 'image',
+                    playlists: playlists,
+                    personalRateOverall: personalRateOverall,
+                    sort: 'm3uExport',
+                    pageNum: Number.isInteger(pageNum) ? pageNum : 1});
+                console.log('START processing: ' + action, searchForm, exportDir, processingOptions);
+
+                promise = tdocManagerModule.exportMediaFiles(searchForm, <MediaExportProcessingOptions & ProcessingOptions> {
+                    ...processingOptions,
+                    exportBasePath: exportDir,
+                    exportBaseFileName: exportName,
+                    directoryProfile: directoryProfile,
+                    fileNameProfile: fileNameProfile,
+                    resolutionProfile: resolutionProfile
+                });
                 break;
             case 'generateTourDocsFromMediaDir':
                 if (importDir === undefined) {