+++ src/backend/modules/tdoc-media-manager.module.ts
@@ -1,11 +1,9 @@
-import {Mapper, utils} from 'js-data';
+import {utils} from 'js-data';
 import {TourDocRecord} from '../shared/tdoc-commons/model/records/tdoc-record';
 import {TourDocDataService} from '../shared/tdoc-commons/services/tdoc-data.service';
 import {TourDocSearchForm} from '../shared/tdoc-commons/model/forms/tdoc-searchform';
 import {TourDocSearchResult} from '../shared/tdoc-commons/model/container/tdoc-searchresult';
 import {BeanUtils} from '@dps/mycms-commons/dist/commons/utils/bean.utils';
-import * as readdirp from 'readdirp';
-import {TourDocAdapterResponseMapper} from '../shared/tdoc-commons/services/tdoc-adapter-response.mapper';
 import {DateUtils} from '@dps/mycms-commons/dist/commons/utils/date.utils';
 import {MediaManagerModule} from '@dps/mycms-server-commons/dist/media-commons/modules/media-manager.module';
 import {SqlConnectionConfig} from './tdoc-dataservice.module';
@@ -19,14 +17,18 @@ import {
     FileSystemDBSyncType
 } from '@dps/mycms-server-commons/dist/backend-commons/modules/cdoc-media-manager.module';
 import {BackendConfigType} from './backend.commons';
+import {TourDocServerPlaylistService} from './tdoc-serverplaylist.service';
+import {TourDocExportService} from './tdoc-export.service';
+import {TourDocMediaFileImportManager} from './tdoc-mediafile-import.service';
 
 export class TourDocMediaManagerModule extends CommonDocMediaManagerModule<TourDocRecord, TourDocSearchForm,
-TourDocSearchResult, TourDocDataService> {
+    TourDocSearchResult, TourDocDataService, TourDocServerPlaylistService, TourDocExportService> {
     private readonly knex;
     private readonly sqlQueryBuilder: SqlQueryBuilder;
 
-    constructor(backendConfig, dataService: TourDocDataService, mediaManager: MediaManagerModule) {
-        super(backendConfig, dataService, mediaManager);
+    constructor(backendConfig, dataService: TourDocDataService, mediaManager: MediaManagerModule,
+                exportService: TourDocExportService, protected mediaFileImportManager: TourDocMediaFileImportManager) {
+        super(backendConfig, dataService, mediaManager, exportService);
         this.sqlQueryBuilder = new SqlQueryBuilder();
         this.knex = this.createKnex(backendConfig);
     }
@@ -42,136 +44,8 @@ TourDocSearchResult, TourDocDataService> {
     }
 
     public generateTourDocRecordsFromMediaDir(baseDir: string): Promise<TourDocRecord[]> {
-        const mapper: Mapper = this.dataService.getMapper('tdoc');
-        const responseMapper = new TourDocAdapterResponseMapper(this.backendConfig);
-        const mediaTypes = this.getFileExtensionToTypeMappings();
-        const fileExtensions = [];
-        // tslint:disable-next-line:forin
-        for (const mediaType in mediaTypes) {
-            fileExtensions.push('*.' + mediaType);
-        }
-        const settings = {
-            root: baseDir,
-            entryType: 'files',
-            // Filter files with js and json extension
-            fileFilter: fileExtensions,
-            // Filter by directory
-            directoryFilter: [ '!.git', '!*modules' ],
-            // Work with files up to 1 subdirectory deep
-            depth: 10
-        };
-
-        const tracks = {};
-        const locations = {};
-        const media = {};
-        const records: TourDocRecord[] = [];
-
-        return new Promise<TourDocRecord[]>((resolve, reject) => {
-            readdirp(settings, function fileCallBack(fileRes) {
-                const path = fileRes['path'].replace(/\\/g, '/');
-                const file = fileRes['name'].replace(/\\/g, '/');
-                const dir = fileRes['parentDir'].replace(/\\/g, '/');
-                const date = fileRes['stat']['mtime'];
-                const extension = file.split('.').splice(-1);
-                const type = mediaTypes[extension];
-                if (type === undefined) {
-                    console.warn('SKIP file - unknown extension', path);
-                    return;
-                }
-
-                if (media[path]) {
-                    return;
-                }
-
-                const values = {
-                    datestart_dt: date,
-                    dateend_dt: date,
-                    dateshow_dt: date
-                };
-                let trackName = dir || file;
-                let locationName = dir || file;
-
-                // normalize trackName
-                let dateStr = '';
-                const matcher = trackName.match(/(\d\d\d\d)(\d\d)(\d\d){0,1}-(.*)?/);
-                if (matcher) {
-                    dateStr = (matcher[3] ? matcher[3] + '.' : '') + matcher[2] + '.' + matcher[1];
-                    locationName = (matcher[4] ? matcher[4] : 'Unbekannt');
-                }
-
-                // normalize locationName
-                const words = locationName.split('-');
-                locationName = '';
-                for (let word of words) {
-                    word = word.substring(0, 1).toUpperCase() + word.substring(1);
-                    locationName = locationName.concat(word).concat(' ');
-                }
-                locationName = locationName.trim();
-                trackName = (locationName + ' ' + dateStr).trim();
-
-                let location: TourDocRecord = locations[locationName];
-                if (!locations[locationName]) {
-                    location = <TourDocRecord>responseMapper.mapResponseDocument(mapper, {
-                        type_s: 'LOCATION',
-                        id: 'LOCATION_' + (records.length + 1),
-                        loc_id_i: records.length + 1,
-                        name_s: locationName,
-                        keywords_txt: 'KW_TODOKEYWORDS',
-                        desc_txt: 'TODODESC'
-                    }, {});
-                    records.push(location);
-                    locations[locationName] = location;
-                }
-                values['loc_id_i'] = location.locId;
-
-                let track: TourDocRecord = tracks[trackName];
-                if (!tracks[trackName]) {
-                    track = <TourDocRecord>responseMapper.mapResponseDocument(mapper, {
-                        type_s: 'TRACK',
-                        id: 'TRACK_' + (records.length + 1),
-                        track_id_i: records.length + 1,
-                        loc_id_i: location.locId,
-                        name_s: trackName,
-                        keywords_txt: 'KW_TODOKEYWORDS',
-                        desc_txt: 'TODODESC',
-                        dateshow_dt: date,
-                        datestart_dt: date,
-                        dateend_dt: date
-                    }, {});
-                    records.push(track);
-                    tracks[trackName] = track;
-                }
-                values['track_id_i'] = track.trackId;
-                if (DateUtils.parseDate(track.datestart).getTime() > DateUtils.parseDate(date).getTime()) {
-                    track.datestart = date;
-                }
-                if (DateUtils.parseDate(track.dateend).getTime() < DateUtils.parseDate(date).getTime()) {
-                    track.dateend = date;
-                }
-
-                if (type === 'IMAGE') {
-                    values['i_fav_url_txt'] = path;
-                    values['image_id_i'] = records.length + 1;
-                }
-                if (type === 'VIDEO') {
-                    values['v_fav_url_txt'] = path;
-                    values['video_id_i'] = records.length + 1;
-                }
-                values['type_s'] = type;
-                values['id'] = type + '_' + (records.length + 1);
-                values['name_s'] = trackName + ' ' + DateUtils.formatDateTime(date);
-                const tdoc = <TourDocRecord>responseMapper.mapResponseDocument(mapper, values, {});
-                media[path] = tdoc;
-                records.push(tdoc);
-            }, function allCallBack(errors) {
-                if (errors) {
-                    errors.forEach(function (err) {
-                        return reject(err);
-                    });
-                }
-                resolve(records);
-            });
-        });
+        return this.mediaFileImportManager.generateTourDocRecordsFromMediaDir(baseDir,
+            this.getFileExtensionToTypeMappings());
     }
 
     public readMetadataForCommonDocRecord(tdoc: TourDocRecord): Promise<{}> {
@@ -199,7 +73,7 @@ TourDocSearchResult, TourDocDataService> {
     }
 
     public findCommonDocRecordsForFileInfo(baseDir: string, fileInfo: FileInfoType,
-                                         additionalMappings: {[key: string]: FileSystemDBSyncType}): Promise<DBFileInfoType[]> {
+                                           additionalMappings: {[key: string]: FileSystemDBSyncType}): Promise<DBFileInfoType[]> {
         return new Promise<DBFileInfoType[]>((resolve, reject) => {
             const createdInSecondsSinceEpoch = Math.round(DateUtils.parseDate(fileInfo.created).getTime() / 1000);
             const lastModInSecondsSinceEpoch = Math.round(DateUtils.parseDate(fileInfo.lastModified).getTime() / 1000);